Enable setting additional emulator flags.
There are (unusual) network configurations where WebRTC does not work.
In these cases you might need to use a turn server.

We update the generation scripts to allow passing in of "extra"
parameters the can be passed on to the emulator. This can be used to
configure the turn server.

- Use argparser to parse parameters.
- Adds documentation
- Fix for file writing flush issues.

Test: python emu_docker.py emu.zip img.zip --extra '-turncfg "curl -s -X POST https://networktraversal.googleapis.com/v1alpha/iceconfig?key=SECRET!"'