Create a python package with a single console entry point
This turns the two scripts into a python package that is ready to be published.
The setup.cfg and setup.py declare all the dependencies as
well as general information about the package.

There is now a single entry shell script (emu-docker) that will be made
available upon install.

Includes some small cleanup to use logging, and wrapping functionality
inside functions.