Add support for interactive selection of system image and emulator
This add the `interactive` command that allows a user to select which
system image and emulator version to use for creating a Docker image.

The system image and emulator image will be downloaded before an image
is created.

Updates the documentation with a quick start section using the
`interactive` command.