Add cleanup before launch
It is possible that the emulator container has data left around. This
can cause trouble where:

- PulseAudio fails to initialize, disabling the video bridge
- Dangling snapshot locks, disabling emulator launch.

Usually this is due to unexpected exits of the emulator executable.

This also includes some cleanup of the Dockerfile to introduce multiple
layers. This makes it faster to rebuild the containers.

Updates the documentation to explain why pre-O images will not work.

Fixes issue #10