Secure the web endpoint
We extend the demo with authentication. This demonstrates how you can
secure the emulator and make it only accessible to authorized users.

Users will login with a username/password and obtain a JWT token. The
JWT token is needed to gain access to the gRPC endpoint of the emulator.

Example run:

```sh

    source ./configure.sh
    emu-docker interactive
    ./create_web_container emulator,hello
    docker-compose -f js/docker/docker-compose.yaml up

```

Point your browser to https://localhost

See README.md for details.