Add health endpoint
The docker container can now report health status. The health status
indicated whether or not the emulator has successfully booted.

Currently this is retrieved by invoking ADB, this is likely going to be
deprecated soon in favor of the gRPC endpoint that is faster and more
accurate, esp with API level >28

Usage:

$ docker inspect <container-id> --format="{{json .State.Health}}"