Prevent early pulse audio termination.
It seems that the switch to the nvidia container, which is based of
ubuntu, caused a different default configuration for pulse audio to be
loaded. In this configuration it is possible for the daemon to exit if
no connections are active.

We now explicitly disable this from the commandline, making sure that
pulse audio sticks around.

Fixes #58