Add support for creating cloud builder releases
This gets the script ready to prepare cloud builder images that
automatically can be pushed to GCE repositories.

This can be used to create self contained commits that can be pushed to
a repository, which in turn will trigger automated deployment of the
images.

This will enable the following workflow:

Emulator Release -> emu-docker cloud-build "(Q.*playstore.*64)"
/my/git/repo -> git push /my/git/repo -> new images that can be pulled.