Generate javascript defintions from distribution
There are a series of changes being made in the emulator gRPC endpoints.
To make sure we can properly handle those we now generate the protobuf
files that match the emulator that is being used.

This will make sure that we do not see strange behavior when running
the web module.