Add support for updated RTC service
The emulator is separating out the WebRTC service. The RTC service will
be hosted on a separate endpoint.

With this change we can create a javascript module that can interact
with both versions.

The Makefile will detect which version of the emulator we should
interact with, and will install the proper web client version.

Test: Both newer and older versions work when creating a webcontainer.