Add support for audio tracks.
This adds support for handling audio tracks that are delivered through
webrtc. This updates the Javascript to be backwards compatible with
older emulators that do not support audio.

In the example audio will be muted, (so we can immediately display the
emulator if possible), and if an audio track becomes available a volume
control will be displayed that can be used to control the audio level.

Note: The volume control does NOT change the volume of the emulated
device itself. For now you will have to manually modify the audio levels
through the settings panel of the android device.

Fixes #135

Test:
 - Able to interact with existing emulator that does not support audio
 - Able to interact with audio enabled emulator.