Improved logcat screen
This adds a react lazylog component to track the logcat messages, it is
based up the react-lazylog component which is added as a submodule.

Includes an update to use the newer gRPC library, and a change to use
polling for logcat, as the streaming rpc call can block the UI thread
for extended periods of time.

This change requires us to migrate to yarn, as it seems that the react
package can only be build with yarn.