Update documentation for TURN usage.
This updates the documentation and adds more details on how to configure
your own TURN server if needed. A TURN server is needed if the emulator
is hidden behind a VPN that does not permit UDP packets to go through.

We modify the launcher to make it easier to pass in the turn
configuration at container launch time.