Add sample of GPS
The newer version of the emulator library has support for gps. This
example demonstrates how you can get the current location and use it to
set the location in the emulator.

Note: Android uses a multitude of ways to derive the location and the
location service might get confused if you "jump" around the earth quickly.