Refactor image creation.
This refactors the docker building into 2 distinct steps:

- Creation of system images.
- Creation of emulator images.

The system images are generally released once a year and do not change
frequently. We are going to publicly host these as a base image on which
we layer the emulator, which we release weekly.

This refactor allows us to host and build emulator images a lot faster
as we can layer on the system image.